requirements.txt
flask
python-dotenv
pymongo 
wheel
requests
flask_cors
FLask-Session
